fixing an off by one bug that addresses #17

+++ demo/lemondetection/execute.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+if [ -f "results.txt" ]; then
+	rm results.txt
+fi
+
 ./main.py
-./combine.py > output.csv
-rm results.txt
+
+if [ -f "results.txt" ]; then
+	./combine.py > output.csv
+	rm results.txt
+fi
